[RENAME] #4 MOVE ErrorRespnse to exception dir
