+++ src/main/java/org/example/bobj/common/exception/GlobalExceptionHandler.java
@@ -4,7 +4,6 @@ package org.example.bobj.common.exception;
 
 import javax.servlet.http.HttpServletRequest;
 import org.example.bobj.common.constants.ErrorCode;
-import org.example.bobj.common.constants.ErrorResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
